refactor: extract duplicated CLI logic to shared helpers
- Move duplicated command actions from cli.ts and quick.ts to cli-helpers.ts
- Extract createTicketAction, updateTicketStatus, listTicketsByStatus, and listAllTicketsAction
- Remove duplicate code between top-level aliases and quick commands
- Improve maintainability by following DRY principles
- All existing functionality preserved and tested

This eliminates ~80 lines of duplicated code and makes the CLI more maintainable
as changes only need to be made in one place.

Functions moved to cli-helpers.ts:
- createTicketAction: shared ticket creation logic
- updateTicketStatus: shared status update logic
- listTicketsByStatus: shared status-based listing
- listAllTicketsAction: shared all-tickets listing
