+++ packages/cli/src/cli.ts
@@ -1,13 +1,7 @@
 import { readFileSync } from 'node:fs'
 import { dirname, join } from 'node:path'
 import { fileURLToPath } from 'node:url'
-import {
-  CreateTicketRequest,
-  GetAllTicketsRequest,
-  SearchTicketsRequest,
-  UpdateTicketStatusRequest,
-} from '@project-manager/core'
-import { CLI, getConfig, isValidTicketStatus } from '@project-manager/shared'
+import { CLI } from '@project-manager/shared'
 import { Command } from 'commander'
 import { configCommand } from './commands/config.js'
 import { createTicketCommand } from './commands/create.js'
@@ -17,14 +11,12 @@ import { createQuickCommands } from './commands/quick.js'
 import { showTicketCommand } from './commands/show.js'
 import { statsCommand } from './commands/stats.js'
 import { updateTicketCommand } from './commands/update.js'
-import { expandPriority, expandType } from './utils/cli-helpers.js'
-import { formatTicketSummaryList } from './utils/output.js'
 import {
-  getCreateTicketUseCase,
-  getGetAllTicketsUseCase,
-  getSearchTicketsUseCase,
-  getUpdateTicketStatusUseCase,
-} from './utils/service-factory.js'
+  createTicketAction,
+  listAllTicketsAction,
+  listTicketsByStatus,
+  updateTicketStatus,
+} from './utils/cli-helpers.js'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const packageJsonPath = join(__dirname, '../package.json')
@@ -67,31 +59,7 @@ function addTopLevelAliases(program: Command): void {
     .option('-p, --priority <priority>', 'Priority (h=high, m=medium, l=low)', 'm')
     .option('-t, --type <type>', 'Type (f=feature, b=bug, t=task)', 't')
     .action(async (title: string, options) => {
-      try {
-        const createTicketUseCase = getCreateTicketUseCase()
-        const config = getConfig()
-
-        // Convert short forms to full values
-        const priority = expandPriority(options.priority)
-        const type = expandType(options.type)
-
-        const request = new CreateTicketRequest(
-          title.trim(),
-          options.description || `Details for: ${title.trim()}`,
-          priority,
-          type,
-          config.defaultPrivacy
-        )
-
-        const response = await createTicketUseCase.execute(request)
-        console.log(`✓ Created ticket ${response.id}: ${response.title}`)
-      } catch (error) {
-        console.error(
-          'Failed to create ticket:',
-          error instanceof Error ? error.message : String(error)
-        )
-        process.exit(1)
-      }
+      await createTicketAction(title, options)
     })
 
   program
@@ -136,72 +104,6 @@ function addTopLevelAliases(program: Command): void {
     .description('List all tickets')
     .option('-c, --compact', 'Compact output format')
     .action(async options => {
-      try {
-        const getAllTicketsUseCase = getGetAllTicketsUseCase()
-        const request = new GetAllTicketsRequest()
-        const response = await getAllTicketsUseCase.execute(request)
-
-        const output = formatTicketSummaryList(response.tickets, {
-          format: options.compact ? 'compact' : 'table',
-        })
-        console.log(output)
-        console.log(`\nTotal: ${response.tickets.length} tickets`)
-      } catch (error) {
-        console.error(
-          'Failed to list tickets:',
-          error instanceof Error ? error.message : String(error)
-        )
-        process.exit(1)
-      }
+      await listAllTicketsAction(options)
     })
-}
-
-/**
- * Helper function to update ticket status
- */
-async function updateTicketStatus(id: string, status: string, action: string) {
-  try {
-    if (!isValidTicketStatus(status)) {
-      console.error(
-        `Invalid status: ${status}. Valid statuses are: pending, in_progress, completed, archived`
-      )
-      process.exit(1)
-    }
-
-    const updateTicketStatusUseCase = getUpdateTicketStatusUseCase()
-    const request = new UpdateTicketStatusRequest(id, status)
-    const response = await updateTicketStatusUseCase.execute(request)
-    console.log(`✓ ${action} ticket ${response.id}: ${response.title}`)
-  } catch (error) {
-    console.error(
-      `Failed to update ticket:`,
-      error instanceof Error ? error.message : String(error)
-    )
-    process.exit(1)
-  }
-}
-
-/**
- * Helper function to list tickets by status
- */
-async function listTicketsByStatus(status: string, format: 'table' | 'json' | 'compact') {
-  try {
-    if (!isValidTicketStatus(status)) {
-      console.error(
-        `Invalid status: ${status}. Valid statuses are: pending, in_progress, completed, archived`
-      )
-      process.exit(1)
-    }
-
-    const searchTicketsUseCase = getSearchTicketsUseCase()
-    const request = new SearchTicketsRequest({ status })
-    const response = await searchTicketsUseCase.execute(request)
-
-    const output = formatTicketSummaryList(response.tickets, { format })
-    console.log(output)
-    console.log(`\n${response.tickets.length} ${status} tickets`)
-  } catch (error) {
-    console.error('Failed to list tickets:', error instanceof Error ? error.message : String(error))
-    process.exit(1)
-  }
 }
